Authorization

+++ app/Http/Controllers/MasyarakatController.php
@@ -9,6 +9,7 @@ class MasyarakatController extends Controller
 {
     public function index()
     {
+        $this->authorize('admin&staff');
         $data["masyarakats"] = User::where('users.level', 'masyarakat')->get();
         $data["title"] = "Masyarakat";
         return view('masyarakat.index', $data);
@@ -16,12 +17,14 @@ class MasyarakatController extends Controller
 
     public function add()
     {
+        $this->authorize('admin&staff');
         $data["title"] = "Tambah Masyarakat";
         return view('masyarakat.add', $data);
     }
 
     public function store(Request $request)
     {
+        $this->authorize('admin&staff');
         $validatedData = $request->validate([
             'nik' => 'required|min:16|unique:users',
             'nama' => 'required',
@@ -40,6 +43,7 @@ class MasyarakatController extends Controller
 
     public function edit(User $masyarakat)
     {
+        $this->authorize('admin&staff');
         $data = User::where('users.username', $masyarakat->username)->get();
         return view('masyarakat.edit', [
             'masyarakat' => $data[0],
@@ -49,6 +53,7 @@ class MasyarakatController extends Controller
 
     public function update(Request $request, User $masyarakat)
     {
+        $this->authorize('admin&staff');
         $rules = [
             'nama' => 'required',
             'telp' => 'required',
@@ -83,6 +88,7 @@ class MasyarakatController extends Controller
 
     public function destroy(User $masyarakat)
     {
+        $this->authorize('admin&staff');
         User::destroy($masyarakat->id);
         return redirect('/masyarakat')->with('success', 'Data Masyarakat Berhasil Dihapus!');
     }
